agricore_project/requirements.txt
django
djangorestframework
drf-nested-routers
django-environ
django-cors-headers
Pillow

channels
channels_redis
redis

celery
django-celery-beat

openai
PyPDF2
python-magic

psycopg2-binary
djangorestframework-simplejwt

gunicorn
dj-database-url
python-dotenv
whitenoise
groq
requests